pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["S4ske <sany1111.ru@mail.ru>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
opencv-python = "^4.10.0.84"
pyjwt = "^2.9.0"
pydantic-settings = "^2.6.0"
passlib = "^1.7.4"
types-passlib = "^1.7.7.20240819"
asyncpg = "^0.30.0"
alembic = "^1.13.3"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-multipart = "^0.0.16"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
redis = "^5.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119
output-format = "grouped"
show-fixes = true
target-version = "py37"
exclude = [".svn", "CVS", ".bzr", ".hg",".git", "__pycache__", ".tox", ".eggs", "*.egg", ".venv", "env", "venv", "build"]

[tool.ruff.lint]
select = ["W", "E", "F", "I", "N", "T20", "Q"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.mccabe]
max-complexity = 6
